Upgraded Sanity Studio to V3
