+++ .dockerignore
@@ -4,4 +4,5 @@ build
 .dockerigore
 .gitignore
 Dockerfile
+studio
 README.md
