+++ .eslintrc.cjs
@@ -1,18 +1,20 @@
 module.exports = {
-  root: true,
-  extends: ["eslint:recommended", "prettier", "plugin:storybook/recommended"],
-  plugins: ['svelte3'],
-  overrides: [{
-    files: ['*.svelte'],
-    processor: 'svelte3/svelte3'
-  }],
-  parserOptions: {
-    sourceType: 'module',
-    ecmaVersion: 2020
-  },
-  env: {
-    browser: true,
-    es2017: true,
-    node: true
-  }
+	root: true,
+	extends: ['eslint:recommended', 'prettier'],
+	plugins: ['svelte3'],
+	overrides: [
+		{
+			files: ['*.svelte'],
+			processor: 'svelte3/svelte3'
+		}
+	],
+	parserOptions: {
+		sourceType: 'module',
+		ecmaVersion: 2020
+	},
+	env: {
+		browser: true,
+		es2017: true,
+		node: true
+	}
 };
