--- miles-conf/.eslintrc
@@ -1,4 +1,0 @@
-{
-  "extends": "@sanity/eslint-config-studio",
-  "eslint.workingDirectories": ["./web", "./studio"]
-}
