--- miles-conf/config/@sanity/data-aspects.json
@@ -1,3 +1,0 @@
-{
-  "listOptions": {}
-}
