--- miles-conf/config/@sanity/default-layout.json
@@ -1,6 +1,0 @@
-{
-  "toolSwitcher": {
-    "order": [],
-    "hidden": []
-  }
-}
