--- miles-conf/config/@sanity/default-login.json
@@ -1,8 +1,0 @@
-{
-  "providers": {
-    "mode": "append",
-    "redirectOnSingle": false,
-    "entries": []
-  },
-  "loginMethod": "dual"
-}
