--- miles-conf/config/@sanity/form-builder.json
@@ -1,5 +1,0 @@
-{
-  "images": {
-    "directUploads": true
-  }
-}
