--- miles-conf/config/@sanity/vision.json
@@ -1,3 +1,0 @@
-{
-  "defaultApiVersion": "2021-10-21"
-}
