--- miles-conf/deskStructure.js
@@ -1,16 +1,0 @@
-// deskStructure.js
-import S from '@sanity/desk-tool/structure-builder'
-
-export default () =>
-  S.list()
-    .title('Base')
-    .items([
-      S.listItem()
-        .title('Site Settings')
-        .child(
-          S.document()
-            .schemaType('siteSettings')
-            .documentId('siteSettings')
-        ),
-        ...S.documentTypeListItems().filter(listItem => !['siteSettings'].includes(listItem.getId()))
-    ])
