--- miles-conf/sanity.json
@@ -1,52 +1,0 @@
-{
-  "root": true,
-  "project": {
-    "name": "miles-conf"
-  },
-  "__experimental_spaces": [
-    {
-      "name": "production",
-      "title": "Prod",
-      "default": true,
-      "api": {
-        "projectId": "mhv8s2ia",
-        "dataset": "production"
-      }
-    },
-    {
-      "name": "test",
-      "title": "Test",
-      "api": {
-        "projectId": "mhv8s2ia",
-        "dataset": "test"
-      }
-    }
-  ],
-  "api": {
-    "projectId": "mhv8s2ia",
-    "dataset": "test"
-  },
-  "plugins": [
-    "@sanity/base",
-    "@sanity/default-layout",
-    "@sanity/default-login",
-    "@sanity/desk-tool"
-  ],
-  "env": {
-    "development": {
-      "plugins": [
-        "@sanity/vision"
-      ]
-    }
-  },
-  "parts": [
-    {
-      "name": "part:@sanity/base/schema",
-      "path": "./schemas/schema"
-    },
-    {
-      "name": "part:@sanity/desk-tool/structure",
-      "path": "./deskStructure.js"
-    }
-  ]
-}
