--- miles-conf/schemas/objects/urls.js
@@ -1,17 +1,0 @@
-export default {
-  name: 'urls',
-  title: 'Urls',
-  type: 'object',
-  fields: [
-    {
-      title: 'URL',
-      name: 'url',
-      type: 'url'
-    },
-    {
-      title: 'Text',
-      name: 'urlText',
-      type: 'string'
-    }
-  ]
-};
