+++ studio/.npmignore
