+++ studio/package.json
@@ -8,27 +8,23 @@
   "license": "UNLICENSED",
   "scripts": {
     "start": "sanity start",
-    "build": "sanity build"
+    "build": "sanity build",
+    "dev": "sanity dev",
+    "lint": "eslint ."
   },
   "keywords": [
     "sanity"
   ],
   "dependencies": {
     "@portabletext/svelte": "^1.0.1",
-    "@sanity/base": "^2.33.0",
-    "@sanity/core": "^2.33.0",
-    "@sanity/default-layout": "^2.33.0",
-    "@sanity/default-login": "^2.33.0",
-    "@sanity/desk-tool": "^2.33.0",
-    "@sanity/eslint-config-studio": "^2.0.0",
-    "@sanity/image-url": "^1.0.1",
-    "@sanity/vision": "2.36.1",
+    "@sanity/icons": "^2.3.1",
+    "@sanity/vision": "^3.12.1",
     "dotenv": "^16.1.4",
-    "eslint": "^8.6.0",
-    "prop-types": "^15.7",
-    "react": "^17.0",
-    "react-dom": "^17.0",
-    "styled-components": "^5.2.0"
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "sanity": "^3.12.1",
+    "styled-components": "^5.3.11"
   },
   "repository": {
     "type": "git",
