+++ studio/sanity.cli.ts
@@ -1,0 +1,8 @@
+import { defineCliConfig } from 'sanity/cli';
+
+export default defineCliConfig({
+	api: {
+		projectId: 'mhv8s2ia',
+		dataset: 'test'
+	}
+});
