+++ studio/schemas/objects/urls.ts
@@ -1,0 +1,17 @@
+export default {
+	name: 'urls',
+	title: 'Urls',
+	type: 'object',
+	fields: [
+		{
+			title: 'URL',
+			name: 'url',
+			type: 'url'
+		},
+		{
+			title: 'Text',
+			name: 'urlText',
+			type: 'string'
+		}
+	]
+};
