+++ studio/schemas/schema.ts
@@ -1,0 +1,26 @@
+import siteSettings from './siteSettings';
+import author from './author';
+import conference from './conference';
+import submission from './submission';
+import externalConference from './externalConference';
+import event from './event';
+
+import portableText from './objects/portableText';
+import performance from './objects/performance';
+import urls from './objects/urls';
+import itinerary from './objects/itinerary';
+import itineraryEvent from './objects/itineraryEvent';
+
+export default [
+  siteSettings,
+  author,
+  conference,
+  externalConference,
+  event,
+  submission,
+  portableText,
+  performance,
+  urls,
+  itinerary,
+  itineraryEvent
+]
